test: refactor change-tracking.test.js in smaller test suites
